refactor Products, minor clean up and fixes
