+++ src/components/Products/ProductItem/index.tsx
@@ -1,24 +1,24 @@
 import { Link } from 'react-router-dom';
-import c from '../Products.module.scss';
+import styles from '../Products.module.scss';
 
 import ImageWithLoader from '../../ImageWithLoader';
-import ItemBottom from './ItemBottom';
-import ItemHeader from './ItemHeader';
+import CardHeader from './CardHeader';
+import CardBottom from './CardBottom';
 import { BaseProduct } from '../../../redux/types';
 
 const ProductItem = ({ product }: { product: BaseProduct }) => {
   return (
     <Link to={String(product.id)} key={product.id} className='product'>
-      <article className={c['product-card']} data-id={product?.id}>
-        <ItemHeader product={product} />
+      <article className={styles['product-card']} data-id={product?.id}>
+        <CardHeader product={product} />
 
         <ImageWithLoader
-          className="rounded-xl"
+          className="rounded-3xl"
           src={product?.image_url.replace('dl=0', 'raw=1')}
-          alt={product.description}
+          alt={'Product Image'}
         />
 
-        <ItemBottom product={product} />
+        <CardBottom product={product} />
       </article>
     </Link>
   );
