+++ src/components/Products/Products.module.scss
@@ -135,6 +135,7 @@ main {
       border-radius: 20px;
       border: 1px solid #eaecf0;
       justify-content: space-between;
+      align-items: center;
       padding: 20px 20px 24px 20px;
       cursor: pointer;
       transition: 0.3s linear;
