+++ src/pages/Catalog/CategoryProducts.tsx
@@ -5,7 +5,7 @@ import {
   useRef,
   useState,
 } from 'react';
-import { Products } from '../../components/Products';
+import Products from '../../components/Products';
 import useFilters from '../../hooks/useFilters';
 import useSorted from '../../hooks/useSorted';
 import { BaseProduct } from '../../redux/types';
@@ -14,7 +14,6 @@ import { getFilteredProducts } from '../../utils/filters';
 type ProductFiltersContext = {
   applyFilters: () => void;
   clearFilters: () => void;
-  triggerForceUpdate: () => void;
 
   isMobileFilterVisible: boolean;
   setIsMobileFilterVisible: React.Dispatch<boolean>;
@@ -39,12 +38,9 @@ export const CategoryProducts = ({
   const [minPrice, setMinPrice] = useState(0);
   const [maxPrice, setMaxPrice] = useState(0);
   const currentPriceRef = useRef(0);
-  const [isForceTriggered, setIsForceTriggered] = useState(false);
 
   const { currentFilterGroups } = useFilters();
   const [isMobileFilterVisible, setIsMobileFilterVisible] = useState(false);
-
-  const triggerForceUpdate = () => setIsForceTriggered(p => !p);
 
   const filterByCurrentCategoryGroups = (products: BaseProduct[]) =>
     getFilteredProducts(products, currentFilterGroups);
@@ -76,10 +72,6 @@ export const CategoryProducts = ({
   };
 
   useEffect(() => {
-    applyFilters();
-  }, [isForceTriggered]);
-
-  useEffect(() => {
     const filteredProducts = filterByCurrentCategoryGroups(products);
 
     if (!loading) {
@@ -102,7 +94,6 @@ export const CategoryProducts = ({
         currentPriceRef,
         applyFilters,
         clearFilters,
-        triggerForceUpdate,
         isMobileFilterVisible,
         setIsMobileFilterVisible,
       }}
